Quick update
Added a better temp asset for the watering can in the inventory, fixed up the journal and some bugs and read through the inventory code to try and understand it. Will start working on improving the inventory tomorrow as we need to get the inventory up to scratch before we get the crafting to work with it.