Fixed shop bug, made cotton in materials section of shop, reshaped UI on canvases to look and work better
Be careful not to disable the shop object in the hierarchy as it was what made the shop go missing. swapped selling washed wool with buying cotton. made all the shop UI tab button fit better and not look so weirds, moved location of item descriptions in dye machine  for consistency.